feat: Add dependabot.yml with scheduled package updates
In this commit, we added a dependabot.yml file to the project. This file configures Dependabot for automated package updates to keep our dependencies up-to-date. 

The settings in dependabot.yml specify the package ecosystem for which updates are to be tracked. Since the value for "package-ecosystem" is left blank (""), Dependabot will automatically detect the package ecosystem based on the project's requirements files.

Additionally, we have scheduled the package updates to occur on a weekly interval.

These settings ensure that our project remains secure and up-to-date with the latest package versions.